fix get_path_id return type

+++ pkg/security/ebpf/c/include/helpers/filesystem.h
@@ -23,7 +23,7 @@ static __attribute__((always_inline)) void bump_path_id(u32 mount_id) {
 #define PATH_ID_LOW_MASK 0xFFFFFF
 #define PATH_ID(high, low) ((u32)((high << 24) | (low & PATH_ID_LOW_MASK)))
 
-static __attribute__((always_inline)) u64 get_path_id(u64 ino, u32 mount_id, int nlink, int invalidate) {
+static __attribute__((always_inline)) u32 get_path_id(u64 ino, u32 mount_id, int nlink, int invalidate) {
     u32 key = mount_id % PATH_ID_MAP_SIZE;
 
     u32 *id = bpf_map_lookup_elem(&path_id, &key);
